backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
alembic==1.13.0
asyncpg==0.30.0
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
pydantic[email]==2.9.0
pydantic-settings==2.5.0
python-dotenv==1.0.1
httpx==0.27.0
openai>=1.30.0
langchain-core>=0.3.0
langchain-openai>=0.3.0
aiofiles>=24.1.0
